requirements.txt
alabaster==0.7.10
attrs==17.4.0
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
Django==2.2
django-braces==1.11.0
django-debug-toolbar==1.8
docutils==0.14
Faker==0.8.8
idna==2.6
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
mixer==6.0.1
more-itertools==4.1.0
packaging==17.1
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.5.1
pytest-cov==2.5.1
pytest-django==3.2.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
South==1.0.2
sqlparse==0.2.3
text-unidecode==1.2
urllib3==1.22
django-extra-views